requirements.txt
pyyaml
opencv-python
numpy
requests